// src/api/generated/interfaces/aboutMeResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * portfolioApi
 * OpenAPI spec version: 1.0
 */

export interface AboutMeResponseDto {
  id?: number;
  /** @nullable */
  nameInPortfolio?: string | null;
  /** @nullable */
  subDescriptions?: string | null;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  descriptionAboutMe?: string | null;
}