// src/api/generated/interfaces/createAboutMeDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * portfolioApi
 * OpenAPI spec version: 1.0
 */

export interface CreateAboutMeDto {
  /**
     * @minLength 1
     * @maxLength 150
     */
  nameInPortfolio: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  subDescriptions: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  profession: string;
  /**
     * @minLength 1
     * @maxLength 1500
     */
  descriptionAboutMe: string;
}