// src/api/generated/interfaces/updateContactDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * portfolioApi
 * OpenAPI spec version: 1.0
 */

export interface UpdateContactDto {
  /**
     * @minLength 1
     * @maxLength 30
     */
  whatsapp: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  email: string;
  /** @minLength 1 */
  linkedin: string;
}